+++ src/lib.rs
@@ -74,6 +74,8 @@ mod platform;
 pub use platform::Enigo;
 
 #[cfg(target_os = "windows")]
+pub use platform::calc_rel_mouse_location;
+#[cfg(target_os = "windows")]
 pub use platform::EXT;
 
 mod keycodes;
@@ -465,7 +467,7 @@ pub struct Settings {
     /// settings for mouse speed and acceleration level. An end user sets
     /// these values using the Mouse application in Control Panel. An
     /// application obtains and sets these values with the
-    /// `windows::Win32::UI::WindowsAndMessaging::SystemParametersInfoA`
+    /// `windows::Win32::UI::WindowsAndMessaging::SystemParametersInfoW`
     /// function. The default value is false.
     pub windows_subject_to_mouse_speed_and_acceleration_level: bool,
 }