+++ src/win/mod.rs
@@ -1,2 +1,2 @@
 mod win_impl;
-pub use win_impl::{Enigo, EXT};
+pub use win_impl::{calc_rel_mouse_location, Enigo, EXT};